fix(budget): Se soluciona el error al guardar un pago con vencimiento 'N/A'
