+++ package.json
@@ -13,35 +13,35 @@
 		"format": "prettier --write ."
 	},
 	"devDependencies": {
-		"@auth/core": "0.18.0",
-		"@auth/sveltekit": "0.3.11",
+		"@auth/core": "0.18.3",
+		"@auth/sveltekit": "0.3.14",
 		"@bdsqqq/try": "2.3.1",
 		"@felte/validator-yup": "1.1.3",
 		"@sveltejs/adapter-auto": "2.1.1",
-		"@sveltejs/kit": "1.27.3",
-		"@trpc/client": "10.43.1",
-		"@trpc/server": "10.43.1",
-		"@typescript-eslint/eslint-plugin": "6.10.0",
-		"@typescript-eslint/parser": "6.10.0",
+		"@sveltejs/kit": "1.27.6",
+		"@trpc/client": "10.44.0",
+		"@trpc/server": "10.44.0",
+		"@typescript-eslint/eslint-plugin": "6.12.0",
+		"@typescript-eslint/parser": "6.12.0",
 		"@vitest/coverage-v8": "0.34.6",
 		"autoprefixer": "10.4.16",
 		"boxicons": "2.1.4",
 		"chart.js": "4.4.0",
-		"daisyui": "3.9.4",
+		"daisyui": "4.4.2",
 		"dayjs": "1.11.10",
 		"delay": "6.0.0",
-		"eslint": "8.53.0",
+		"eslint": "8.54.0",
 		"eslint-config-prettier": "9.0.0",
-		"eslint-plugin-svelte": "2.35.0",
+		"eslint-plugin-svelte": "2.35.1",
 		"felte": "1.2.12",
 		"postcss": "8.4.31",
-		"prettier": "3.0.3",
-		"prettier-plugin-svelte": "3.0.3",
-		"superjson": "2.2.0",
-		"svelte": "4.2.2",
+		"prettier": "3.1.0",
+		"prettier-plugin-svelte": "3.1.1",
+		"superjson": "2.2.1",
+		"svelte": "4.2.7",
 		"svelte-chartjs": "3.1.2",
-		"svelte-check": "3.5.2",
-		"svelte-sonner": "0.3.0",
+		"svelte-check": "3.6.0",
+		"svelte-sonner": "0.3.3",
 		"tailwindcss": "3.3.5",
 		"trpc-sveltekit": "3.5.22",
 		"tslib": "2.6.2",
@@ -54,7 +54,7 @@
 	"type": "module",
 	"overrides": {
 		"svelte-chartjs": {
-			"svelte": "4.2.2"
+			"svelte": "4.2.7"
 		}
 	},
 	"dependencies": {
