+++ src/lib/components/budget/DetailsLogs.svelte
@@ -37,15 +37,15 @@
 	}
 </script>
 
-<section class="md:col-span-2 lg:col-span-4 my-3">
-	<Logs title="Movimientos" className="!max-h-44" background="bg-base-100" on:change={handleChange}>
+<section class="sm:col-span-2 lg:col-span-5 my-3">
+	<Logs title="Movimientos" on:change={handleChange}>
 		<svelte:fragment slot="columns">
 			<th class="px-4">Monto ($)</th>
 		</svelte:fragment>
 		<svelte:fragment slot="body">
 			{#if loading}
 				{#each awaitLoad as _}
-					<tr class="animate-pulse">
+					<tr class="skeleton">
 						<td><div class="bg-slate-400 h-5" /></td>
 						<td><div class="bg-slate-400 h-5" /></td>
 						<td><div class="bg-slate-400 h-5" /></td>
