+++ src/lib/components/shared/Table.svelte
@@ -1,8 +1,13 @@
 <script lang="ts">
 	export let className = ''
+	export function scrollBottom() {
+		if (ref) ref.scroll({ top: ref.scrollHeight, behavior: 'smooth' })
+	}
+
+	let ref: HTMLElement
 </script>
 
-<div class={`overflow-x-auto max-h-80 ${className}`}>
+<div bind:this={ref} class={`overflow-x-auto max-h-80 ${className}`}>
 	<table class="table table-pin-rows">
 		<thead>
 			<slot name="head" />
