+++ src/lib/components/shared/tabs/TabItem.svelte
@@ -4,7 +4,7 @@
 	export let text: string
 </script>
 
-<div class="basis-full tooltip tooltip-top" data-tip={text}>
+<div class="tab basis-full tooltip tooltip-top" data-tip={text}>
 	<a class="tab w-full gap-x-2 text-lg" {href} class:tab-active={active}>
 		<slot />
 		<span class="hidden sm:block text-sm">{text}</span>
