+++ src/routes/+error.svelte
@@ -6,7 +6,7 @@
 	<title>{$page.status} Error!</title>
 </svelte:head>
 
-<div class="fixed inset-0 flex flex-col justify-center items-center bg-primary-focus">
+<div class="fixed inset-0 flex flex-col justify-center items-center bg-neutral">
 	{#if $page.status === 404}
 		<svg
 			xmlns="http://www.w3.org/2000/svg"
@@ -28,7 +28,7 @@
 						d="M561 249c.727-.663 1.853-1.095 1.988-2.175.133-1.068-.874-2.846-1.192-3.825l-.796 6M399.614 454.552c-1.89-.008.154-3.94.457-4.842.596 1.019 2.022 4.853-.457 4.842m15.75-49.73c.745-.445 2.152-.963 2.558-1.805.47-.975-.495-.966-1.342-1.025-1-.07-2.315-.135-3.293.055-1.003.196-1.432.426-2.371-.167-.533-.335-1.038-1.168-1.65-1.29-.636-.126-1.37.546-1.92.801-1.92.887-3.957.313-5.993.41-.769.037-1.407.22-2.12-.111-.912-.424-1.12-.988-2.241-.913-1.817.123-3.208.677-5.07.181-1.155-.307-2.744-1.526-3.987-1.15-1.154.349-1.757 2.164-2.67 2.933-.306-1.583-.473-3.332-1.686-4.52-1.359-1.33-3.56-1.807-5.134-2.9-3.554-2.466-6.128-6.015-8.837-9.321-1.72 2.596-2.943 4.975-4.092 7.89-1.097 2.782-2.163 5.575-3.26 8.357-2.21 5.6-4.567 11.146-6.988 16.66-4.915 11.194-10.433 22.082-15.268 33.319 3.307.08 6.537 1.706 9.386 3.265.745.408.777.569 1.256 1.213.231.31.546 1.07.966 1.174.246.06 1.031-.376 1.42-.32.582.085.99.463 1.43.812a167.954 167.954 0 0 1 8.204 6.871c1.213 1.087 2.036 2.114 2.86 3.515.906 1.545 1.887 2.527 3.269 3.654 2.23 1.818 3.519 5.01 5.078 7.443.895 1.398 1.768 2.817 2.453 4.332.62 1.37.836 3.016 1.963 4.09 2.383 2.265 6.111.964 8.999 1.339 1.136.147 1.41.468 2.432.01.012-.403.165-.547.46-.432.337.086.644.239.92.457.752.409 1.468.902 2.197 1.351l1.276-3.145c-.478-.01-2.205.2-2.474-.521-.226-.605.722-.774 1.079-1.148.396-.416.628-1.769 1.4-1.741.632.023.597.577.588 1.118 1.362.029 1.35-1.158 1.581-2.169.157-.68.377-1.351.554-2.028.117-.456.882-2.325.689-2.717-.486-.977-3.607-.917-1.617-2.107.634-.38 2.122-.26 2.614-.796.391-.43.426-2.19.542-2.779-.928.065-2.044.626-2.613-.2-.822-1.194.378-3.017.985-4.014.784.617 1.274 1.292 2.168.578.96-.765 1.834-2.194 2.191-3.296-1.036-.031-.851.175-1.378-.633-.3-.462-.386-1.263-.535-1.78-.304-1.063-.92-1.873.362-2.157.802-.177 1.002.008 1.235.77.307 1.004.196 2.725.316 3.798 1.255-1.583 2.542-3.287 2.07-5.434-.799.795-.88.58-1.844-.038-.962-.617-1.909-1.257-2.862-1.89 1.206-1.093 3.119-.034 4.307-.932.808-.61.395-2.483.45-3.325.068-1.06.528-2.131.465-3.192-.066-1.147-1.381-1.65-.684-2.702.49-.737 1.625-.626 1.837-1.73.214-1.107-.906-1.774-1.74-2.257 1.234-.401 3.604-.828 2.49-2.48-.527-.78-1.165-.848-1.154-1.956.01-.801.538-1.667.88-2.364.461 1.058 1.142 3.541 2.578 3.457 1.197-.07 2.26-1.145 1.915-2.228-1.061-3.338-4.08-3.581-7.017-4.209.549-2.023 1.997-1.906 3.46-.879.867.607 1.409 1.119 2.512.493 1.15-.652.594-.895.087-1.832-1.035-1.906.398-2.999 1.8-4.11.808-.64 1.2-1.44 1.89-2.135.707-.71 1.851-.867 2.743-1.284-1.068-.522-2.036-1.067-3.225-.66-.79.272-1.368.802-2.27.344-2.019-1.022.603-1.283.872-2.127.152-.48-.366-1.454-.17-2.135.2-.696.633-1.287 1.019-1.887.82-1.275 1.717-2.563 1.228-4.153.988 1.137 1.32 1.387 1.399 2.986.07 1.42.079 2.846.364 4.244.061-2.061 1.953-3.435 2.271-5.418.098-.618-.404-2.133.185-2.463.46-.258 1.401.4 1.888.403-.339-1.883-.395-3.866-1.032-5.685-.623-1.782-2.244-2.466-3.604-3.633"
 					/>
 				</g>
-				<g fill="currentColor">
+				<g fill="#aaa">
 					<path
 						d="M933.692 204.51c-2.507.524-4.924 1.334-7.45 1.815-2.994.568-5.992 1.09-9.01 1.528-2.025.294-3.858.756-5.189 2.449-2.477 3.153-4.97 8.59-1.559 11.936 1.604 1.572 4.416 1.077 6.477 1.731 2.296.73 4.435 1.539 6.842 1.853 11.221 1.457 22.382 4.17 33.36 6.87 5.01 1.232 10.008 2.936 14.797 4.858 2.452.984 4.848 2.13 7.04 3.613 1.875 1.267 4.014 2.516 5.524 4.218-6.004 1.02-12.745 1.139-18.75.12-3.011-.512-6.042-.369-9.08-.599-3.044-.23-6.088-.446-9.129-.714-2.95-.26-5.881-.587-8.811-1.026-1.085-.162-2.183-.323-3.281-.18-.638.082-1.041.276-1.624.144-.488-.111-1.012-.494-1.502-.658-2.248-.75-4.647-.606-6.978-.74-1.747-.1-3.611-.625-5.36-.238-3.074.678-4.476 4.996-5.18 7.643-.442 1.668-.09 3.741.45 5.39.62 1.89 1.758 1.855 3.282 2.713 2.178 1.227 4.089 2.568 6.438 3.525 1.967.803 3.98 1.5 5.927 2.354 4.287 1.877 8.935 3.506 12.973 5.875 2.19 1.286 4.262 2.57 6.558 3.684a44.463 44.463 0 0 1 5.949 3.457c3.852 2.672 7.204 5.994 10.062 9.703-1.326 0-2.409.172-3.678-.278-1.559-.55-3.016-1.359-4.595-1.857-3.008-.95-5.957-1.845-8.931-2.928-2.971-1.083-6.015-2.156-9.049-3.052-1.555-.46-3.126-.664-4.703-1.001-1.391-.299-2.626-.923-3.94-1.45-6.31-2.534-13.063-3.753-19.59-5.582-6.357-1.78-12.64-3.63-19.13-4.883-3.02-.584-6.047-1.135-9.097-1.554-2.985-.41-5.585.003-8.47.768-3.036.806-5.79 1.552-8.689 2.828-2.072.912-5.143 1.71-6.759 3.284-.966.942-1.597 1.874-2.827 2.562-.935.523-1.95.888-2.954 1.252-2.276.825-4.431 1.91-6.693 2.764-1.424.538-2.9.962-4.424 1.06-1.457.093-1.967.446-3.274 1.084-2.044.994-4.285 1.461-6.44 2.14-2.146.677-4.588 2.466-6.912 1.651-2.258-.793-3.704-3.701-5.05-5.506-3.548-4.745-6.597-9.445-8.57-15.058-2.092-5.953-4.088-11.939-6.113-17.915-1.052-3.108-2.11-6.213-3.183-9.314-.597-1.72-1.833-3.572-.485-5.161 1.564-1.844 2.12-4.748 3.737-6.812 1.607-2.051 3.419-3.937 5.122-5.907.995-1.15 1.966-2.324 2.852-3.561.726-1.014 1.264-2.257 2.07-3.198.732-.855 1.809-1.358 2.612-2.148.873-.859 1.65-1.877 2.451-2.805 2.102-2.436 4.101-5.018 6.343-7.326 1.85-1.906 3.964-3.602 5.865-5.48 4.117-4.064 8.49-7.858 12.757-11.76 4.155-3.8 8.405-7.478 12.449-11.368 3.8-3.657 8.36-6.613 12.62-9.67 4.666-3.35 9.14-6.884 14.165-9.708 4.807-2.703 9.646-3.517 15.012-4.318-.53 1.042-.908 2.145-1.61 3.087-1.039 1.39-2.398 2.63-3.573 3.907a1007.678 1007.678 0 0 1-7.206 7.758c-4.703 5.009-9.629 9.815-14.193 14.952-3.637 4.093-6.464 8.27-7.966 13.589-.792 2.804-1.186 5.812.515 8.389 1.302 1.97 3.442 1.817 5.6 1.624 5.56-.498 11.076-2.39 16.58-3.373 6.118-1.093 12.374-1.361 18.55-2.02 3.1-.33 6.161-1.183 9.261-1.417 1.398-.105 2.778.18 4.175.064 1.39-.116 2.783-.455 4.152-.709 2.879-.534 5.494-.838 8.408-.45 2.897.386 5.666.084 8.564-.154 6.12-.505 12.393.156 18.48.816 5.827.632 11.606 1.334 17.297 2.724-5.96 1.516-11.909 3.07-17.903 4.453-6.119 1.41-12.437 2.36-18.444 4.18-2.614.792-5.3.906-7.962 1.463ZM461.873 330.8c.653-2.372 1.314-4.92 2.144-7.287.191-.547.486-.972.795-1.377-.563.938-.96 1.87-1.34 3.128-.555 1.839-1.066 3.69-1.599 5.535Zm-1.306 2.425Zm-.477 1.174Zm1.654-3.147Zm364.847-22.126c-2.377 1.514-4.69 2.471-6.637 4.57-4.39 4.728-9.118 9.203-13.342 14.076-1.698 1.956-2.555 4.38-4.452 6.202-2.156 2.069-4.786 3.68-7.225 5.385-5.418 3.787-10.782 7.88-16.499 11.202-1.282.745-2.58 1.517-3.902 2.184-.87.438-1.966.674-2.728 1.294-.726.59-1.019 1.467-1.81 2.041-.582.424-1.282.77-1.892 1.157a857.973 857.973 0 0 1-15.14 9.362c-4.963 3-9.927 6.04-14.786 9.207-2.582 1.685-5.021 3.577-7.615 5.238-2.113 1.353-4.32 2.556-6.392 3.972-2.197 1.502-4.248 3.04-6.653 4.217-2.326 1.138-4.745 2.08-7.06 3.24-1.365.682-2.6 1.512-3.9 2.297-.85.512-1.563.438-2.488.754-2.696.923-5.113 2.521-7.863 3.373-11.426 3.54-22.474-1.51-32.433-6.881-5.02-2.708-9.59-6.174-14.399-9.225-2.102-1.333-4.726-2.302-6.65-3.869-.967-.788-1.34-1.618-2.472-2.24-1.025-.56-1.959-1.124-2.883-1.844-4.559-3.548-8.818-7.191-12.887-11.3-2.221-2.244-4.442-4.495-6.608-6.79-1.697-1.8-3.02-3.985-4.586-5.915-3.77-4.639-7.426-9.346-11.06-14.09-1.764-2.303-3.509-4.622-5.278-6.922-1.18-1.534-2.959-3.049-3.364-5.035-.374-1.826.698-3.574 1.58-5.08 1.327-2.269 2.594-4.523 3.484-7.007 1.339-3.736 3.382-10.893-.706-13.573-1.503 2.497-3.404 4.718-5.025 7.134-1.63 2.434-2.333 5.51-4.29 7.678-2.057 2.276-4.796 3.927-6.953 6.146-2.15 2.212-4.117 4.594-6.2 6.87-2.18 2.381-4.495 4.633-6.79 6.903-2.158 2.135-4.622 3.85-7.052 5.665-2.543 1.9-5.163 3.591-7.897 5.202-1.18.696-2.396 1.554-3.69 2.02-1.379.494-2.589.792-3.915 1.478-5.397 2.793-10.84 5.473-16.319 8.097-1.47.705-3.02 1.684-4.6 2.076-.607.15-.954.013-1.622.227-.703.223-1.403.62-2.078.917-2.812 1.242-5.683 2.199-8.596 3.175-6.06 2.03-12.215 3.884-18.209 6.1-5.61 2.077-11.068 4.322-16.85 5.93-5.697 1.585-11.714 1.85-17.579 2.424-1.576.155-3.426-.007-4.94.433-1.36.395-2.112.685-3.593.566-2.62-.21-5.067.288-7.666.687 3.957-7.789 6.012-16.66 8.877-24.91.757-2.182 1.677-4.218 2.307-6.437.561-1.984 1.378-3.888 2.362-5.698 1.2-2.208 1.568-4.763 2.37-7.132-.624 1.862-.962 3.844-1.554 5.71.822-1.568 1.706-3.105 2.49-4.694.82-1.66.81-2.648.82-4.434 2.126.245 2.206-1.941 2.534-3.5.498-2.37 1.024-4.735 1.534-7.102.258-.546.17-1.444.306-2.039.162-.704.432-1.379.672-2.057 1.855-5.256 3.894-10.423 5.24-15.843 1.486-5.976 2.786-12.007 4.122-18.018.71-3.191 1.65-6.443 1.96-9.702.148-1.562.944-6.6-1.268-7.302-1.709-.543-3.54 4.666-3.986 5.813-1.084 2.787-2.108 5.596-3.14 8.403-.923 2.51-1.638 5.045-2.667 7.518-1.04 2.5-1.405 5.104-2.327 7.64-.541 1.49-1.412 2.587-2.008 4.012-.434 1.041-.594 2.171-.95 3.238-.74 2.21-1.885 4.372-2.718 6.572-1.066 2.81-2.379 5.517-3.548 8.285-1.07 2.532-1.68 5.2-2.684 7.738-.966 2.438-2.464 4.615-3.425 7.075-1.056 2.703-2.016 5.442-3.06 8.15-2.154 5.587-4.336 11.131-6.025 16.88-.422 1.439-1.142 2.852-1.493 4.133-.319 1.162-.887 2.082-1.344 3.238-.818 2.073-1.16 4.405-1.575 6.589l.124-.707c-.862.036-1.786.224-2.386-.547-.685-.877-.494-2.579-.622-3.632-1.014.133-.618.59-.507 1.507l.169 1.417.426 3.544c.323-.19 1.003-1.292 1.28-1.323.423-.047.884 1.102 1.08 1.67a4.61 4.61 0 0 1-.268.63c-.268.525-.524.89-.729 1.46-.38 1.063-.668 2.145-1.209 3.147a7.96 7.96 0 0 1-.925 1.386c-.27.316-1.15.84-1.298 1.173-.369.833.194 2.912.141 3.882-.067 1.261-.269 2.611-.82 3.763-.232.483-.653.776-.87 1.193-.306.588-.32.892-.391 1.654-.25 2.715-1.29 5.259-2.23 7.791-1.053 2.844-1.876 5.765-2.834 8.64-.905 2.722-2.064 5.339-3.021 8.04-1.116 3.143-2.11 6.33-3.154 9.496-.781 2.371-1.977 4.478-2.815 6.81-.86 2.387-2.07 4.533-2.488 7.014-.463 2.747-1.91 5.496-2.953 8.067-.872 2.148-2.677 4.431-3.036 6.719-.168 1.074.098 2.077-.256 3.144-.307.926-.894 1.716-1.285 2.602-.537 1.22-.137 2.348-.507 3.594-.321 1.085-1.054 1.847-1.646 2.773-1.043 1.634-.69 3.53-1.762 5.234-1.512 2.405-2.047 5.1-2.858 7.832-.4 1.341-.694 2.723-1.052 4.075-.142.539-.213 1.61-.499 2.107-.357.625-.968.38-1.278.74-.267.312-.326 1.32-.444 1.72-.374 1.27-.91 2.502-1.425 3.722-1.11 2.621-2.384 5.172-3.344 7.853-1.004 2.804-2.012 5.608-3.018 8.412-.478 1.335-.943 2.673-1.46 3.992-.506 1.296-1.037 1.692-1.886 2.683-.836.974-1.182 2.34-1.965 3.373-.736.969-1.232 1.68-1.554 2.886-.606 2.269-1.053 4.548-2 6.716-2.394 5.481-5.378 10.695-7.78 16.183-2.518 5.751-4.908 11.562-7.088 17.45-2.047 5.526-3.355 11.256-5.448 16.763-2.072 5.451-3.852 10.985-5.894 16.452-2.12 5.678-4.25 11.346-6.148 17.101-.685 2.076-1.145 4.14-1.589 6.23.138-.94.271-1.863.43-2.729-2.021 3.381-3.878 7.058-2.991 11.11.552-.78.971-1.663 1.307-2.6-.128.483-.27.955-.467 1.374-.566 1.203-1.098 2.12-1.483 3.453-.851 2.95-1.715 5.899-2.552 8.853-.404 1.423-.781 2.86-1.255 4.263-.421 1.247-1.336 2.329-1.722 3.546-.433 1.362-.285 3.03-.619 4.448-.292 1.247-.72 2.464-1.36 3.577-.497.865-1.208 1.62-1.602 2.544-.48 1.126-.19 2.243-.32 3.428-.165 1.474-.693 2.878-.894 4.345-.194 1.413.465 2.29.716 3.64.263 1.411-.27 2.343-.69 3.632-.359 1.096-.566 2.264-.824 3.386-.111.48-.19 1.985-.53 2.305-.249.233-1.678.27-2.037.333-1.346.236-2.69.496-4.063.496-2.88 0-5.832-.419-8.693-.726-5.792-.621-11.617-1.537-17.17-3.333-5.267-1.704-10.576-4.194-15.463-6.793-4.927-2.62-9.298-6.208-13.416-9.945-2.235-2.029-4.363-4.173-6.492-6.312-1.024-1.028-2.046-2.057-3.083-3.072-.826-.811-1.897-1.584-1.98-2.837-.094-1.412.926-3.03 1.423-4.304.444-1.139.881-2.28 1.313-3.423 1.056-2.793 2.093-5.593 3.132-8.391 2.153-5.796 4.292-11.596 6.447-17.391 2.17-5.83 4.396-11.533 7.017-17.172 2.509-5.4 4.792-10.844 6.877-16.422 1.969-5.267 4.312-10.344 6.545-15.5 1.198-2.77 2.332-5.57 3.25-8.445.768-2.408 1.338-4.68 2.507-6.94.752-1.455 1.365-2.878 1.822-4.454.395-1.363.617-2.53 1.5-3.635 1.863-2.329 2.902-4.959 3.86-7.764 1.867-5.463 4.03-10.986 6.644-16.138 2.647-5.222 5.578-10.232 7.99-15.572 2.428-5.37 5.731-10.2 8.003-15.656 4.686-11.254 11.147-21.616 15.897-32.84 4.678-11.056 10.25-21.678 15.028-32.69a491.488 491.488 0 0 0 7.18-17.389c1.085-2.79 2.16-5.584 3.283-8.358 1.014-2.504 2.768-4.424 3.49-7.01.739-2.647 1.921-4.685 3.175-7.07 1.294-2.459 1.963-5.428 2.985-8.02a665.997 665.997 0 0 1 6.963-16.893c2.458-5.76 5.082-11.44 7.634-17.157.964-2.162 1.504-4.486 2.497-6.633.494-1.068 1.496-1.974 1.938-2.996.206-.474.157-1.134.327-1.642.23-.68.656-1.362.972-2.007 2.56-5.24 5.395-10.28 7.73-15.63 2.386-5.47 4.826-10.934 6.924-16.522 2.073-5.522 4.695-10.729 6.478-16.369.849-2.689 1.696-5.39 2.346-8.135.567-2.393 1.634-5.923 1.093-8.381-.333-1.513-2.091-4.595-3.873-3.2-1.639 1.284-2.625 3.827-3.52 5.634-1.346 2.717-2.774 5.4-3.927 8.208-.92 2.242-1.749 4.404-3.023 6.476-3.07 4.996-5.513 10.227-8.278 15.395-.625 1.17-1.361 2.138-2.212 3.142-.825.974-.591 1.44-.929 2.674-.604 2.21-2.44 3.99-3.371 6.079-1.239 2.78-2.312 5.398-3.862 8.037a79.51 79.51 0 0 0-3.667 6.962c-2.521 5.462-4.944 10.971-7.387 16.47-1.296 2.915-2.554 5.853-3.976 8.71-.625 1.254-1.236 2.527-1.914 3.754-.44.798-1.238 1.449-1.61 2.26-.423.92.042 1.696-.32 2.593-.417 1.041-.872 1.465-1.78.856-2.029-1.356-2.433-4.261-3.521-6.333-2.57-4.896-5.018-9.498-5.978-15.037-.531-3.07-.32-6.197-.3-9.295.02-2.75 1.652-4.617 3.088-6.89 3.371-5.345 6.748-10.685 10.178-15.99 1.51-2.334 2.825-4.788 4.37-7.094 1.493-2.228 3.484-3.874 5.123-5.942.914-1.152 1.583-2.416 2.56-3.526.82-.936 1.605-1.907 2.276-2.957 1.279-1.999 2.152-4.162 3.556-6.1 3.313-4.577 6.684-9.128 10.174-13.571 1.86-2.37 4.052-3.838 6.354-5.726 2.254-1.849 4.271-3.958 6.349-6 4.001-3.934 7.774-8.042 12.224-11.46a214.241 214.241 0 0 1 14.085-9.959c2.376-1.537 5.326-2.696 7.236-4.776.788-.858 1.605-1.66 2.615-2.258.97-.574 2.1-.727 3.042-1.252 2.671-1.487 5.128-3.927 7.557-5.789 2.42-1.851 4.838-3.704 7.257-5.555 1.017-.78 2.935-1.7 3.677-2.704.417-.563.367-1.349.839-1.942 1.068-1.341 3.18-1.467 4.678-2.044 2.878-1.107 5.813-1.703 8.833-2.287 2.662-.515 5.203-1.657 7.759-2.552 5.593-1.96 11.39-3.138 17.288-3.67-2.652 7.822-5.663 15.53-7.691 23.55-.821 3.25-.96 7.043-2.33 10.105-.492 1.103-1.81 2.153-1.151 3.398.488.927 1.86 1.93 2.736 2.499 1.817 1.182 4.06-.129 5.845-.843 3.01-1.203 5.927-2.75 9.153-3.304 3.35-.573 6.164-3.746 9.642-3.694 1.454.022 2.634.717 3.723 1.62 1.121.932 2.39 1.68 3.477 2.662 2.388 2.156 5.167 4.007 8.188 5.202 1.048.414 1.942.2 2.912 1.003.724.599 1.299 1.461 2.268 1.683-4.593 4.905-9.047 9.764-12.935 15.268-1.905 2.696-3.324 5.331-4.366 8.469-.493 1.483-.948 2.616-.002 4.05.523.794 1.358 1.447 2.336 1.296 2.363-.364 4.462-3.555 5.773-5.303 1.52-2.03 3.191-3.91 4.775-5.889 3.29-4.11 6.21-8.424 10.307-11.805 1.22-1.006 2.454-2.148 3.806-2.975 1.078-.66 1.839-.156 3.033-.366 2.99-.526 5.675-2.09 8.347-3.442a633.822 633.822 0 0 1 8.705-4.347c1.526-.744 3.773-2.316 5.565-1.564 2.004.84 3.137 3.977 4.236 5.655 1.055 1.613 2.006 3.015 3.737 3.932 2.84 1.507 8.359 3.336 10.847.437 3.523-4.105 3.165-10.828 2.861-15.862-.187-3.086-.565-6.196-.328-9.288.208-2.72.528-5.433.796-8.15 5.007 4.107 9.738 8.275 13.887 13.265 2.207 2.656 4.495 5.241 6.774 7.836 1.14 1.3 2.167 2.436 2.946 3.966.671 1.318 1.627 2.458 2.624 3.539 2.086 2.258 3.51 5.256 5.1 7.89a476.77 476.77 0 0 1 4.549 7.68c2.913 5.038 5.642 9.996 7.87 15.375 1.158 2.796 2.732 5.335 3.916 8.118 1.171 2.76 2.33 5.528 3.458 8.308.651 1.604 1.433 3.223 1.94 4.877.352 1.15.041 2.605.463 3.673.405 1.024 1.376 1.922 1.805 3.025.434 1.116.69 2.293 1.133 3.406 2.171 5.45 5.148 10.647 7.58 15.99 2.522 5.544 5.04 11.097 8.058 16.396 3.062 5.376 6.922 10.307 10.615 15.258 1.934 2.592 3.935 5.145 5.789 7.793.668.953 1.263 2.047 2.155 2.817.843.728 1.908 1.013 2.888 1.499.541-1.44 1.278-2.384 1.046-3.98-.196-1.359-.702-2.691-1.123-3.989-.813-2.51-2.945-4.754-2.41-7.536.499-2.595 2.622-4.651 4.23-6.611 2.125-2.59 4.26-5.173 6.398-7.753 7.183-8.662 14.811-16.962 22.25-25.424 3.787-4.307 7.5-8.686 11.416-12.877 1.755-1.88 3.793-3.408 5.664-5.16 1.534-1.435 2.544-3.247 3.99-4.742 1.073-1.106 2.347-1.913 3.438-3.031l2.66-2.725a875.565 875.565 0 0 0 5.833-6.01c3.868-4.03 7.481-8.281 11.291-12.363 7.357-7.882 15.14-15.45 22.82-23.011 2.218 6.551 4.052 13.38 5.641 20.11.391 1.657.553 3.38 1.14 4.974.578 1.566 1.334 3.073 1.942 4.63 1.35 3.46 2.383 7.15 4.064 10.448 3.191 6.26 5.997 13.522 10.594 18.911 2.288 2.68 5.421 4.977 8.087 7.281l8.613 7.45c-2.584 1.945-5.147 3.897-7.876 5.637Zm-237.763 41.592c-.237-.554-.464-.851-.733-1.05.2.119.376.306.605.749.493.95.796 2.047 1.148 3.056-.318-.927-.635-1.854-1.02-2.755Zm-69.39 49.636c1.984 2.22 4.085 4.338 6.093 6.536 1.852 2.027 3.338 4.25 5.001 6.426a264.29 264.29 0 0 0 11.223 13.7c3.745 4.28 7.789 8.315 11.364 12.742 3.574 4.427 6.273 9.175 8.95 14.166 1.241 2.313 2.654 4.535 4.33 6.557.784.944 1.837 1.771 2.33 2.907.547 1.258.714 2.778 1.026 4.108 2.785 11.894 5.636 23.778 8.32 35.695 1.302 5.776 2.173 11.63 3.431 17.414 1.274 5.855 3.183 11.542 4.387 17.417 2.44 11.898 4.108 23.954 6.728 35.813 2.553 11.547 6.984 22.523 9.391 34.12.339 1.628.827 3.114.45 4.76-.314 1.37-.555 2.572-.579 3.992-.042 2.443-.277 4.366-1.693 6.45-2.79 4.11-6.687 7.686-11.2 9.738-2.421 1.1-4.724 2.58-7.339 3.2-1.258.296-2.48.1-3.727.285-1.266.19-2.31 1.069-3.474 1.546-4.656 1.914-10.6 1.34-15.543 1.344-5.447.005-10.916-.36-15.945-2.64a27.004 27.004 0 0 1-1.218-.59c-.679-.104-.799-.41-.359-.915.002-.551.315-1.18.34-1.763.051-1.24-.319-2.431-.63-3.615-.737-2.81-1.43-5.632-2.074-8.466-2.665-11.75-3.928-23.681-5.567-35.601-.844-6.148-2.074-12.24-3.133-18.354-1.015-5.869-1.625-11.761-2.393-17.664-.765-5.88-1.582-11.756-2.375-17.632-.432-3.218-1.365-6.378-1.775-9.581-.31-2.427.332-4.718-.218-7.151-1.293-5.727-1.346-11.665-2.805-17.355-1.504-5.863-3.678-11.625-5.702-17.325-1.047-2.952-2.09-5.906-3.138-8.858-.618-1.746-1.326-3.62-3.121-4.456-.996-.463-1.981-.411-2.876-1.102-1.098-.848-2.042-2.003-3.02-2.986-2.104-2.118-4.193-4.25-6.305-6.36-8.217-8.212-16.18-16.821-25.25-24.095-4.215-3.38-8.124-7.193-12.793-9.946-5.004-2.95-9.938-6.014-14.944-8.96a55.41 55.41 0 0 0-7.661-3.76c-2.35-.94-.915-3.187-.237-5.007 1.049-2.814 1.958-5.67 2.835-8.543.663-2.17 1.025-4.98 3.246-6.157 1.453-.77 2.088.127 3.481.562.99.308 2.03.356 3.057.311 2.326-.102 5.005-1.145 7.297-.782 1.535.243 2.298.303 3.85.021 1.382-.25 2.838-.255 4.239-.359 2.587-.193 5.248-.225 7.803-.674 5.594-.984 11.26-2.273 16.625-4.152 5.744-2.012 11.218-4.806 16.722-7.378.37 2.534.733 5.11 1.67 7.51.747 1.92 1.984 4.489 3.817 5.507 4.014 2.23 6.114 6.07 9.089 9.4Zm83.955 241.454c1.743-.948 2.075.383 3.11 1.812 2.826 3.906 5.672 7.798 8.526 11.685 2.777 3.784 5.564 7.564 8.344 11.347 2.607 3.545 5.487 6.928 7.881 10.623 2.46 3.791 4.41 7.954 6.298 12.05 1.604 3.487 3.935 7.117 2.25 11.001-1.564 3.606-6.016 5.129-9.255 6.844-4.092 2.167-7.32 3.7-11.95 2.051-8.811-3.137-17.442-7.213-25.79-11.41-7.65-3.848-13.438-9.533-19.78-15.117-3.33-2.932-6.78-5.63-9.895-8.806-3.238-3.304-6.323-6.757-9.584-10.038-5.073-5.11-10.393-9.627-9.517-17.537 4.84 1.621 9.601 3.657 14.63 4.63 5.298 1.024 10.702 1.65 16.103 1.37 2.328-.122 4.457-.757 6.727-1.208 2.216-.44 4.474.019 6.556-1.07 1.842-.963 2.984-2.106 5.091-2.608 2.144-.51 3.953-1.212 5.775-2.475 1.494-1.036 2.88-2.273 4.48-3.144ZM450.63 431.929c.05.01.097.028.142.056a.754.754 0 0 0-.142-.056Zm-5.641-3.187c.052.033.105.071.16.11-.053-.039-.107-.074-.16-.11ZM341.233 708.115a440.206 440.206 0 0 1 6.696 12.35 238.179 238.179 0 0 1 3.124 6.248c.939 1.934 1.78 3.432 1.892 5.564-4.525-.505-9.12-3.186-12.963-5.555-4.041-2.49-7.465-5.574-10.894-8.826-3.346-3.17-6.76-6.307-10.343-9.21-1.653-1.34-3.38-2.49-5.207-3.574-.558-.33-1.123-.67-1.577-1.143-.556-.58-.757-1.592-1.432-2.03-1.379-.893-3.534.44-4.769 1.044-1.635.8-3.209 1.757-4.58 2.962-1.57 1.382-2.623 3.153-3.975 4.728-1.269 1.477-2.552 2.888-3.693 4.47-2.585 3.58-5.395 7.388-8.688 10.35-1.502 1.35-3.395 2.423-4.683 3.946a30.908 30.908 0 0 1-4.01 3.963c-3.267 2.688-6.1 5.376-8.904 8.546-3.09 3.495-6.605 6.626-9.942 9.88-.512-2.06.403-3.784 1.312-5.561 1.058-2.074 1.807-4.314 2.69-6.466a152.331 152.331 0 0 1 6.448-13.744c2.214-4.164 4.48-8.296 6.626-12.497 1.78-3.485 4.534-7.786 3.684-11.858-.332-1.588-1.196-3.985-2.911-4.516-1.71-.53-3.428.971-4.806 1.787-4.038 2.388-8.22 4.269-12.545 6.073-4.313 1.8-8.663 3.55-12.814 5.705-4.027 2.089-7.827 4.588-11.855 6.669-1.992 1.028-4.101 1.79-6.07 2.865-1.898 1.037-3.734 2.18-5.622 3.232-3.73 2.08-7.616 3.535-11.857 4.145 3.025-3.898 7.005-6.757 10.86-9.76 4.497-3.499 8.966-7.024 13.609-10.33 4.173-2.97 8.418-5.83 12.502-8.925 1.91-1.447 3.795-2.926 5.62-4.477 1.534-1.304 3.183-2.608 3.553-4.708.333-1.89-.53-4.07-2.133-5.164-1.6-1.091-3.635-.843-5.43-.587-5.047.718-9.967 2.26-14.852 3.68-1.124.326-2.275.86-3.4 1.123-1.135.264-2.473.153-3.635.212-2.18.11-4.356.225-6.523.502-2.264.29-4.575.207-6.825.568-2.048.327-3.472 1.436-5.566 1.386-9.298-.225-17.928 4.944-27.127 4.944a2228.587 2228.587 0 0 1 16.498-9.583c2.527-1.454 5.157-2.748 7.6-4.346 2.702-1.767 5.184-3.774 7.975-5.426 9.985-5.91 21.095-10.229 30.203-17.51 2.209-1.766 4.316-3.663 6.448-5.52.482-.42 1.117-1.202 1.691-1.473.914-.428.62-.548 1.423.274.883.904 1.495 2.009 2.227 3.035 2.9 4.066 6.461 7.634 10.482 10.593 4.674 3.438 10.06 5.525 15.297 7.94 5.28 2.433 10.49 5.026 16.021 6.857 5.577 1.847 11.482 2.794 17.305 3.481 6.164.73 11.913.316 18.03-.446.356 4.601.822 8.498 2.991 12.666 2.047 3.933 4.648 7.58 6.824 11.447Zm23.178-350.5c.45.099 1.024.136 1.454.294-.373-.094-.848-.097-1.125-.175-.115-.032-.22-.078-.329-.119Zm1.65 2.656c.177-.438.44-.928.485-1.378-.027.445-.267.94-.484 1.378Zm-52.668 77.857c-6.861 4.082-13.01 9.214-19.655 13.618-3.318 2.198-6.778 4.175-10.326 5.977-1.711.87-3.436 1.733-5.22 2.447-1.548.618-3.036 1.615-4.62 2.084-1.723.51-3.489.202-5.233-.004-1.514-.178-2.986-.082-4.232-1.084-2.25-1.807-3.708-5.42-5.045-7.879-1.912-3.516-4.148-7.183-5.517-10.942-1.276-3.51-2.959-6.72-3.625-10.397 1.179-1.956 1.075-3.992.086-5.987-.455-.918-1.042-1.763-1.58-2.631-.609-.982-.582-1.21-1.676-1.228l3.152 9.79c-1.037-3.21-2.439-6.496-3.152-9.79-.624-2.874-.863-5.661-1.889-8.463-.524-1.424-1.35-2.763-1.76-4.22-.337-1.199-.329-2.466-.489-3.694-.446-3.425-1.326-6.817-1.867-10.235-1.03-6.523-1.691-13.118-2.94-19.604-1.285-6.672-2.94-13.303-4.396-19.941-.555-2.534-1.549-4.705.925-6.433 2.368-1.652 5.002-3.03 7.503-4.468 2.374-1.366 4.735-2.746 7.262-3.808 2.145-.902 4.208-1.762 6.159-3.04 1.214-.795 2.457-.79 3.83-1.233 1.414-.457 2.81-.987 4.21-1.482 2.574-.91 5.143-1.828 7.714-2.745 1.406-.501 4.294-2.343 5.611-.947.683.721.762 2.23.955 3.131.316 1.484.616 2.972.908 4.461 1.128 5.77 2.21 11.56 3.105 17.372.879 5.702 1.379 11.459 2.265 17.158.217 1.396.923 3.12.783 4.525-.102 1.035-.821 1.582-1.53 2.256a67.505 67.505 0 0 0-5.092 5.398c-1.77 2.097-3.462 4.266-5.375 6.238-1.864 1.922-3.484 3.65-4.828 5.994-1.128 1.968-2.636 3.814-2.113 6.17.409 1.83.942 3.652 3.116 2.899 2.515-.87 4.689-3.337 6.653-5.05 1.93-1.682 3.836-3.39 5.803-5.028 2.173-1.812 4.602-3.34 6.703-5.226 1.825-1.638 3.387-3.543 5.119-5.274 4.153-4.156 8.622-7.881 13.04-11.739 2.037-1.779 3.66-3.954 5.506-5.918 1.82-1.934 3.797-3.714 5.708-5.556 4.146-3.995 8.459-8.019 12.321-12.29 1.85-2.046 3.835-3.957 5.699-5.99.769-.839 1.339-1.944 2.07-2.743.707-.773 1.795-1.27 2.587-1.993 3.464-3.16 5.394-7.56 8.81-10.782.909 1.988 1.551 4.017 2.218 6.1.615 1.92 1.078 4.162 2.157 5.88.503.798 1.357 1.378 1.792 2.21.416.799.306 1.747.6 2.591.563 1.61 1.724 3.306 2.59 4.778.967 1.644 2.336 3.14 3.63 4.53.334.36.67.68 1.039.956-.151-.084-.3-.169-.457-.248 1.03 1.753 1.961 2.56 3.637 3.712l-.14.291c-.866 1.874-1.84 3.61-2.521 5.57-1.177 3.387-2.905 6.555-4.017 9.967-.598 1.835-.83 3.731-1.701 5.47-.766 1.53-1.762 2.932-2.511 4.472-1.75 3.603-2.891 7.555-4.27 11.31-1.425 3.879-3.07 7.629-4.831 11.364a232.483 232.483 0 0 0-4.928 11.264c-.697 1.71-1.248 3.62-2.349 5.126-1.19 1.63-2.906 2.052-4.714 2.698-7.162 2.56-14.14 6.403-20.667 10.285ZM227.57 328.122c-4.055-3.431-9.715-2.464-14.695-2.485-5.785-.025-11.543 1.113-17.308 1.5-5.762.388-11.552.664-17.329.544-5.152-.108-10.286-.591-15.335-1.641-2.5-.52-4.944-.773-7.482-.993-1.59-.136-3.06-.619-3.86-2.108-1.054-1.966-.78-4.955-.855-7.12-.074-2.122-.38-4.816.465-6.815.832-1.972 2.683-2.94 4.573-3.664 2.32-.89 4.643-1.772 6.963-2.66 9.4-3.601 19.07-7.722 29.196-8.587 2.812-.24 5.635-.265 8.454-.35 2.421-.072 4.66-.66 7.053-.91 4.995-.521 9.88-.635 14.808-1.68 1.354-.288 2.206.433 3.582.653 1.074.172 2.165.14 3.243.062 2.228-.16 4.356-.434 6.593-.243 2.63.223 5.276.361 7.884.768 2.343.365 4.612.998 6.975 1.266 2.553.29 5.136.4 7.65.964 1.27.283 3.206.557 3.954 1.779.99 1.615.247 4.717.15 6.501-.136 2.513-.29 5.03-.264 7.548.027 2.734.436 5.378.964 8.055.38 1.928 1.195 4.191.76 6.18-.24 1.094-1.039 1.611-2.173 1.435-.785-.122-1.43-.89-2.155-.983-1.409-.181-.04.871-.847 1.775-.647.725-1.719.522-2.58.761-.926.258-1.782.711-2.657 1.097-1.983.873-4.041 1.475-5.994 2.438-4.342 2.14-8.468 4.68-12.6 7.191-.691-2.058-1.185-5.176-2.576-6.877-1.197-1.463-3.13-2.192-4.557-3.4Zm284.537-130.14c.931-2.626 2.157-5.192 2.933-7.867.277-.956.277-2.314 1.256-2.852.813-.448 2.286-.165 3.14-.159 1.58.013 3.16-.005 4.739-.012.79-.003 1.932-.182 2.398.645.48.855-.29 2.423-.306 3.375-.024 1.444.318 2.857.684 4.243 1.563 5.907 3.184 11.743 4.98 17.594-1.79-.805-3.182-2.722-5.226-2.939-1.29-.137-2.282.635-3.471.831-1.017.168-2.03.227-3.041.463-4.544 1.057-8.675 3.413-13.29 4.241.214-1.69.282-3.233.801-4.869.452-1.421.953-2.83 1.439-4.241.971-2.823 1.968-5.638 2.964-8.453Zm14.994-39.632c2.183-.966 4.366-3.768 2.322-6.081-1.102-1.247-3.893-.593-5.334-.296-2.022.417-9.786 5.481-10.741 2.038-.628-2.26.496-5.666 1.523-7.651.65-1.258 1.524-2.486 2.57-3.45 1.12-1.036 1.583-.72 2.962-.567 2.276.254 4.391-.45 5.887 1.77.291.43.35 1.017.96 1.107.47.069.75-.514 1.224-.515.767-.002.853.53 1.152 1.125.783 1.555 1.934 4.611 4.265 3.338 2.356-1.285 2.497-5.45 2.5-7.694.005-3.482.24-6.96.43-10.435.148-2.751.39-5.261 2.395-7.335 1.892-1.955 3.358-4.675 5.657-6.173 2.595-1.692 5.972-1.502 8.946-1.73 3.437-.264 6.964-.642 10.403-.289 3.52.361 6.868.527 10.41.367 3.461-.156 6.917-.371 10.382-.46 3.29-.082 6.728.194 9.985-.16 1.324-.143 3.09-.625 3.965.687.995 1.49.912 4.19 1.191 5.907-3.58-.66-9.423-2.312-10.797 2.748-.416 1.53.872 2.897 2 3.787 1.65 1.3 3.958 1.264 5.96 1.544 1.117.157 2.92.128 3.574 1.25.63 1.077.3 3.087.374 4.271.305 4.924 1.008 9.812 1.52 14.716.489 4.687.858 9.497.15 14.182-.362 2.402-1.346 4.66-1.52 7.08-.109 1.499.125 3.199-1.243 4.19-1.588 1.149-4.321 1.155-6.185 1.565-2.46.542-4.894 1.194-7.314 1.893-3.835 1.109-8.514 1.662-9.897 5.951-1.476 4.578-2.564 9.297-3.991 13.898-.726 2.334-1.738 4.449-2.764 6.662-.397.854-1.02 1.846-.938 2.828.086 1.036.81 2.046 1.21 2.97.875-.456 1.792-.71 2.538-1.382.781-.7 1.405-1.688 2.033-2.527 1.305-1.743 2.233-3.239 4.104-4.372.071.922.29 1.713-.152 2.494-.312.555-1.034 1.12-1.445 1.635-1.151 1.445-2.26 2.928-3.509 4.29-2.486 2.714-5.494 4.95-8.08 7.583-2.593 2.64-5.046 5.738-8.208 7.715-3.047 1.905-6.304 3.48-9.474 5.167 2.786-6.33.858-12.702-2.457-18.394-1.543-2.65-3.42-5.081-5.157-7.602-1.375-1.995-4.032-4.672-3.968-7.236.035-1.364.965-2.541 1.636-3.668.744-1.249 1.32-2.58 1.93-3.9 1.237-2.672 2.85-5.498 3.284-8.44.38-2.575-.202-6.143-1.943-8.205-.7-.827-1.726-.747-2.635-1.309-.893-.553-1.569-1.38-2.302-2.116-.832-.835-1.852-1.412-2.681-2.237-.88-.877-1.498-1.99-2.135-3.047-1.51-2.503-3.057-4.985-4.572-7.487Zm9.576-55.636c2.62.046 4.01 1.524 5.23 3.632-.817 1.283-3.006 5.153-4.812 3.61-.334-.286-.49-.903-.896-1.235-.382-.313-.861-.477-1.304-.674-1-.442-2.133-.87-2.773-1.831-1.508-2.262 2.972-3.527 4.555-3.502Zm51.18 98.181c.47-.127 1.921.825 2.383 1.029.85.377 1.908.746 2.66 1.314.753.569 1.03 1.418 1.546 2.18.156 5.698.515 11.439.472 17.154l-.474-17.159c-.334 1.57-.914 3.012-.912 4.634 0 1.643.25 3.335.404 4.967.29 3.109.637 6.213.955 9.319a84.633 84.633 0 0 1-.277 5.36c-2.273-3.162-4.503-6.363-6.782-9.52-2.038-2.82-4.7-2.984-7.917-3.242l6.312-12.744c.241-.487 1.21-3.178 1.63-3.292Zm371.464 9.659c3.394-.939 6.86-2.243 10.358-2.705 3.31-.436 6.251-1.875 9.447-2.812 2.816-.824 5.666-1.53 8.44-2.494 2.108-.731 4.384-1.454 6.044-3.014 1.719-1.617 2.647-4.13 2.328-6.48-.374-2.745-2.4-3.08-4.71-3.867-11.765-4.006-23.672-6.393-35.951-7.976-6.32-.816-12.711-1.83-19.097-1.7-3.309.07-6.58.252-9.888.04-2.994-.191-6.043.93-8.988.346-3.01-.596-5.822.207-8.856.619-3.224.438-6.483.566-9.733.716-2.953.136-6.09.016-8.962.79-2.809.754-4.984 1.296-7.852.361 2.998-3.316 5.985-6.642 8.994-9.948 6.177-6.785 12.59-14.077 16.327-22.56 1.602-3.637 2.38-8.878-.292-12.263-2.376-3.01-7.566-2.39-10.876-1.896-4.256.633-8.32 2.15-11.982 4.393-3.53 2.162-7.431 3.928-10.643 6.536-3.156 2.562-6.791 4.257-9.982 6.77-3.29 2.593-6.604 5.103-9.932 7.638-7.19 5.476-13.162 12.434-19.968 18.353-3.617 3.146-6.984 6.534-10.218 10.068-1.533 1.675-3.122 3.326-4.432 5.186-1.096 1.556-2.049 3.213-3.204 4.73-1.386 1.814-3.126 3.327-4.586 5.085-1.34 1.618-2.573 3.322-3.762 5.053-1.251 1.822-2.453 3.676-3.695 5.503-.652.96-1.275 2.596-2.585 2.428-1.85-.237-2.874-1.826-4.886-.73-.68.368-1.104.997-1.67 1.496-.795.699-1.707.997-2.632 1.46-1.888.947-3.484 2.353-5.175 3.593-1.793 1.314-3.633 2.398-5.294 3.898-6.978 6.305-13.418 13.06-20.128 19.628-1.622 1.587-3.322 3.094-4.899 4.728-1.254 1.301-2.236 2.846-3.497 4.126-.728.74-1.605 1.038-2.374 1.681-.717.602-1.317 1.55-1.929 2.252-1.473 1.692-2.93 3.398-4.444 5.055-2.965 3.25-6.241 6.078-9.092 9.45-2.911 3.445-6.192 6.595-9.168 9.997-6.262 7.161-12.143 14.474-17.9 22.043-2.856 3.755-6.069 7.243-8.793 11.094-2.589 3.659-4.778 7.58-6.732 11.61-.552 1.137-1.26 4.525-3.29 3.327-.727-.429-1.214-1.692-1.664-2.393-.608-.951-1.213-1.903-1.816-2.856-2.625-4.154-5.434-8.308-7.509-12.771-3.993-8.592-7.966-17.113-11.251-26.008-1.661-4.5-3.557-8.888-5.448-13.295-1.873-4.368-3.293-8.897-5.12-13.28-3.636-8.719-8.414-16.967-13.168-25.112-4.356-7.464-9.025-15.164-15.084-21.384-3.204-3.288-5.723-7.059-9.11-10.21-1.715-1.597-3.487-3.133-5.2-4.733-1.239-1.155-2.293-2.87-3.824-3.647-1.083-.549-2.24-.43-3.41-.382-.908.036-1.915-.23-2.79-.151-.788.073-1.106.758-1.868.91-.778.158-1.292-.203-1.904-.625-3.868-2.669-9.155-3.823-13.57-5.423 2.794-1.73 6.22-1.793 9.34-2.522 2.638-.618 5.102-1.706 7.27-3.34 4.779-3.605 5.575-9.303 6.44-14.879.995-6.407 1.766-13.237.913-19.69a1117.625 1117.625 0 0 0-2.665-18.993c-.46-3.07-1.047-6.066-1.962-9.036-.744-2.42-1.703-4.917-1.617-7.484.042-1.268.085-3.226 1.045-4.144.707-.673 1.747-1.033 2.5-1.696 1.762-1.551 3.179-3.444 4.263-5.522.886-1.696 2.316-4.183 1.47-6.145-.92-2.128-3.845-2.52-5.808-3.028-5.812-1.51-11.681-1.266-17.585-2.097-2.836-.399-5.802-1.226-8.65-1.191-2.806.033-5.608.079-8.417.045-6.12-.073-12.261-.251-18.36.379a103.81 103.81 0 0 1-4.834.388c-1.31.076-2.25.742-3.454.95-.733.126-1.77-.27-2.506-.328a14.315 14.315 0 0 0-3.73.178c-2.594.486-4.607 2.139-7.177 2.656-2.002.403-3.534 1.494-4.557 3.268-2.597 4.495-2.738 10.696-2.965 15.73-.267 5.963.177 11.95-.136 17.904-.076 1.44-.23 2.022-1.754 1.872-2.137-.211-4.01-.783-6.17-.213-4.117 1.087-6.697 4.833-8.303 8.536-1.686 3.882-2.589 8.263-2.303 12.501.25 3.728 1.624 8.55 5.733 9.628 2.205.58 4.417-.34 6.499-.994 1.957-.616 3.741-2.38 4.804.094 2.152 5.013 4.32 9.548 7.761 13.865-3.228-.22-6.45-.532-9.68-.694-2.123-.107-5.131-.7-7.163.127-1.194.487-1.246 1.326-2.099 2.07-.94.82-2.762.47-3.926.434-3.484-.107-6.681-.741-10.207-.375-2.596.269-4.94 1.555-7.518 2.044-3.136.593-6.252 1.837-9.211 3.014-2.784 1.11-5.516 2.297-8.38 3.192-2.922.913-5.843 1.852-8.63 3.132-4.96 2.28-9.171 5.964-13.61 9.093-1.167.823-2.44 1.53-3.53 2.458-.892.76-1.482 1.764-2.343 2.55-1.378 1.26-2.987 1.197-4.48 1.864-1.528.683-3.003 2.052-4.39 2.999-1.253.857-2.508 1.713-3.763 2.57-2.384 1.625-4.741 3.559-7.317 4.859-4.784 2.416-8.538 6.244-12.426 9.885-4.196 3.93-8.825 7.354-12.866 11.442-4.03 4.079-7.713 8.407-10.955 13.136-3.38 4.928-6.232 10.152-9.435 15.188-1.465 2.303-3.32 4.174-5.028 6.282a36.68 36.68 0 0 0-2.396 3.304c-.652 1.016-1.112 2.312-2.084 3.056-1.946 1.486-3.888 2.363-5.457 4.35-1.756 2.227-3.27 4.688-4.759 7.098-2.983 4.84-5.783 9.844-8.44 14.87-.786 1.488-.777 2.577-.91 4.2-.1 1.259-.514 2.46-.747 3.696-.381 2.013-.146 4.128-.667 6.126-.34 1.31-1.182 1.482-2.245 2.22-1 .693-1.903 1.512-2.78 2.352-1.785 1.713-3.52 3.403-5.47 4.934-1.982 1.555-4.035 3.057-5.759 4.908-1.79 1.925-3.365 3.772-5.424 5.444-4.375 3.55-8.504 7.48-12.801 11.141-4.462 3.802-8.53 7.974-12.635 12.15-4.243 4.316-8.733 8.363-13.164 12.483-.348-4.89-.635-9.784-1.042-14.67-.44-5.274-1.627-10.416-2.677-15.596-.938-4.625-1.562-9.575-3.132-14.035-1.077-3.056-3.183-5.567-6.44-6.32-1.63-.377-4.099.395-4.888-1.422-.785-1.806-.86-4.554-1.177-6.494-1.398-8.565 1.38-18.125-2.198-26.32-1.553-3.556-4.44-5.44-7.875-6.964-4.498-1.995-9.467-1.499-14.237-2.488-2.177-.452-3.948-1.642-6.208-1.743-2.281-.102-4.566-.027-6.846.06-2.381.09-4.672-.388-7.037-.38-2.06.006-3.68 1.375-5.701 1.054-2.267-.358-4.116-.916-6.406-.258-2.028.582-3.962.756-6.069.814-2.595.074-5.191-.171-7.785-.046-2.384.114-4.658.838-7.02 1.128-4.995.612-9.809 1.181-14.683 2.524-4.828 1.33-9.413 3.247-14.1 4.985-7.513 2.786-16.134 4.92-18.203 13.775-1.146 4.91-1.375 10.199-.758 15.196.498 4.043 1.415 8.52 3.794 11.921 2.542 3.639 6.759 4.741 10.797 5.995 4.628 1.434 9.363 1.654 14.1 2.513 4.914.889 9.827 1.6 14.837 1.465 4.969-.134 9.829-1.427 14.783-1.848 5.094-.432 10.17-1.03 15.26-1.474 3.424-.3 9.968-1.27 11.684 2.616.542 1.229 1.32 2.697.327 3.796-1.183 1.308-2.228 2.648-3.295 4.103-1.448 1.973-2.53 3.863-2.96 6.297-.424 2.387-.49 4.822-.524 7.24-.138 9.976 2.43 20.088 4.865 29.744 2.532 10.047 5.352 20.04 8.882 29.784 3.363 9.285 5.862 18.95 9.851 27.994 2.003 4.54 4.62 8.568 7.376 12.672 2.318 3.451 4.438 7.627 7.852 10.131 2.863 2.102 6.526 3.135 10.067 2.852 4.245-.338 8.278-2.914 11.995-4.795 8.951-4.532 17.221-10.657 25.251-16.654 3.855-2.878 7.35-6.18 11.454-8.72 4.008-2.484 8.052-4.71 12.567-6.017-3.764 7.933-7.185 15.992-11.186 23.814-3.948 7.721-7.215 15.72-10.768 23.62-1.461 3.249-3.614 6.06-5.053 9.296-1.648 3.71-3.302 7.419-4.944 11.131-3.004 6.782-6.202 13.528-8.918 20.424-.617 1.565-.34 3.457-1.14 4.832-.806 1.387-1.687 2.599-2.3 4.105-1.317 3.241-2.362 6.507-3.862 9.682-3.322 7.029-5.432 14.59-8.612 21.692-3.188 7.12-6.064 14.315-8.906 21.58-1.393 3.561-2.96 7.056-4.26 10.653-1.2 3.318-2.327 6.714-4.285 9.682-4.297 6.516-7.628 13.552-10.863 20.659-1.526 3.352-2.872 6.682-3.102 10.411-.208 3.357 1.872 6.076 3.551 8.778-3.057 1.548-5.873 3.548-8.939 5.108-3.29 1.675-6.724 2.989-10.146 4.36-3.46 1.386-6.913 2.791-10.36 4.213-3.038 1.253-5.847 2.793-8.793 4.227-2.787 1.357-5.328 3.147-7.965 4.762-2.852 1.747-5.747 3.426-8.6 5.173-5.592 3.425-11.035 6.859-16.905 9.804-3.06 1.535-6.137 3.038-9.2 4.568-2.517 1.258-5.282 2.364-7.12 4.59-2.718 3.29-2.481 8.584.099 11.933 1.704 2.21 5.888 2.004 8.431 2.076 1.734.049 3.497-.046 5.2-.395 1.408-.287 2.603-1.037 3.976-1.387 1.555-.395 3.294-.31 4.891-.524a69.451 69.451 0 0 0 4.658-.805c3.347-.683 6.665-1.507 10.007-2.217 3.256-.691 6.52-1.362 9.713-2.311 1.325-.394 2.724-1.15 4.088-1.12 1.482.03 2.778.133 4.27-.133 1.524-.272 3.019-.699 4.554-.904 1.427-.19 2.868.039 4.28-.2 1.592-.27 3.07-.973 4.612-1.427 1.508-.446 2.897-.307 4.448-.307-3.822 2.956-7.819 5.678-11.713 8.537-3.934 2.886-7.605 6.075-11.387 9.149-3.062 2.485-6.89 4.876-8.775 8.452-1.516 2.879-1.798 6.446-1.188 9.614.766 3.974 3.707 4.478 7.282 4.726 4.62.32 8.182-1.637 12.215-3.624 7.898-3.891 15.53-8.385 22.994-13.045 3.693-2.306 7.774-4.044 11.697-5.92 2.075-.992 4.153-1.978 6.232-2.961 1.7-.805 3.924-2.408 5.8-2.224-5.163 10.696-10.567 21.277-15.116 32.259-1.907 4.604-3.425 9.364-3.8 14.36-.263 3.505-.08 10.446 4.298 11.606 1.954.518 3.714-.778 5.028-2.07 1.873-1.843 3.598-3.86 5.388-5.785a174.275 174.275 0 0 1 10.98-10.79c3.872-3.49 7.742-6.898 11.218-10.792 3.313-3.711 7.453-6.523 10.608-10.38 3.24-3.963 6.146-8.186 9.409-12.13 2.147-2.596 3.872-6.661 7.556-7.238.549 3.064 3.191 5.788 5.039 8.292 4.078 5.53 9.003 10.06 14.097 14.63 4.607 4.135 9.559 7.601 15.37 9.841 3.825 1.475 11.824 4.892 14.827.251 1.428-2.206.674-6.512.14-8.927-.741-3.354-2.368-6.496-3.837-9.58-2.987-6.275-6.297-12.298-9.657-18.373-1.687-3.053-3.342-6.126-5.162-9.101-1.58-2.58-3.775-5.068-4.552-8.047-.322-1.225-.52-2.888-.124-4.125.42-1.315.97-1.073 2.148-1.241 2.585-.369 1.806-5.141 1.467-7.002-.301-1.657-.922-3.257-1.517-4.827-.504-1.331-1.626-3.145-1.563-4.586.058-1.364 1.381-2.753 1.917-3.977.623-1.423 1.031-2.911 1.472-4.398 2.019-6.817 4.664-13.401 6.344-20.321 1.672-6.887 3.448-13.624 5.9-20.285 2.542-6.907 5.315-13.745 7.628-20.735 2.245-6.781 4.455-13.547 6.921-20.255 4.948-13.454 10.22-26.809 16.053-39.906 5.755-12.924 12.293-25.503 17.267-38.762 2.393-6.377 4.756-12.71 7.01-19.137a374.847 374.847 0 0 1 7.541-19.708c2.76-6.654 5.036-13.523 7.396-20.326 2.322-6.696 4.309-13.619 7.173-20.1.737.802 1.415 1.378 1.944 2.325.528.943.976 1.827 1.706 2.644 1.312 1.465 2.98 2.49 4.57 3.616.996.704 1.821 1.545 2.96 1.7l.09.012c-.205.012-.41.022-.64.069 1.45 1.006 2.9 2.01 4.353 3.02a13.09 13.09 0 0 0-.383-.817c.471.82.962 1.594 1.98 1.769a2.36 2.36 0 0 1-.245.029c.688.509 3.131 2.941 2.916.318.094.425-.065.746.113 1.172.189.449.627.86.956 1.22.69.75 1.494 1.387 2.318 1.98 1.935 1.39 4.021 2.557 5.9 4.028.543.425 2.387 2.565 3.023 2.564l3.557 3.101c-.453-1.589-.909-3.176-1.364-4.764l.99 3.455c.383 1.339.962 1.824 1.968 2.723 1.765 1.578 3.515 3.177 5.347 4.68 3.342 2.738 7.178 4.862 10.757 7.273 1.947 1.31 3.856 2.656 5.709 4.094 1.52 1.18 3.378 1.597 4.89 2.741 1.901 1.443 3.588 3.322 5.238 5.039 1.48 1.538 3.226 3.089 4.248 4.984 1.157 2.149 1.176 4.95 1.536 7.329.365 2.4.726 4.801 1.076 7.204.721 4.93 1.403 9.866 1.995 14.813.582 4.864 1.296 9.724 1.719 14.604.428 4.938.32 9.882.876 14.816 1.112 9.896 2.373 19.984 4.492 29.72 2.136 9.81 3.504 19.76 5.358 29.632.954 5.077 1.963 10.144 2.958 15.213.489 2.493 1.016 4.98 1.51 7.471.407 2.06.195 4.13.557 6.207.192 1.112.463 2.225.463 3.36.001 1.104-.691 2.07-.73 3.058-.055 1.392 1.22 3.588 1.992 4.697 1.166 1.674 2.78 3.188 4.133 4.724 1.363 1.547 1.911 2.102 1.427 4.192-2.078 8.963 3.246 16.472 8.93 22.809 6.258 6.976 12.517 13.7 19.696 19.739 3.76 3.164 7.3 6.579 10.899 9.922 3.45 3.205 7.172 5.707 11.204 8.151 4.086 2.478 8.417 4.284 12.652 6.461 4.176 2.147 8.234 4.184 12.675 5.748 4.223 1.487 8.5 2.753 12.553 4.643 3.2 1.491 7.26 1.427 10.491.048 2.175-.93 4.042-2.623 5.984-3.947a86.777 86.777 0 0 1 5.571-3.524c3.146-1.813 6.222-3.924 7.565-7.464 1.59-4.192.215-9.226-.694-13.436-.93-4.304-2.85-8.314-5.15-12.044-2.398-3.884-5.38-7.245-8.132-10.865-2.847-3.744-5.383-7.685-8.36-11.34-2.949-3.616-6.038-7.064-9.357-10.34-3.312-3.267-6.205-6.91-9.26-10.412-.736-.843-1.861-1.749-2-2.944-.12-1.041.596-1.793.94-2.71.666-1.778.469-3.732.503-5.59.017-.905.002-1.835.453-2.655.42-.765 1.2-1.07 1.498-1.892.265-.732.168-1.658.315-2.423.187-.97.64-1.677.95-2.598.773-2.303-.186-4.988-.62-7.272-.41-2.157-.797-4.318-1.195-6.476-.86-4.667-1.502-9.345-2.968-13.874-2.98-9.193-5.894-18.338-6.092-28.1-.1-4.889-1.463-9.573-2.593-14.3a2041.085 2041.085 0 0 0-3.652-15.026c-2.369-9.584-5.21-19.028-7.637-28.595-2.419-9.532-4.62-19.127-7.498-28.533-2.908-9.507-6.95-18.432-11.274-27.366-2.08-4.295-4.356-8.003-7.416-11.672-3.173-3.803-6.427-7.54-9.687-11.27-6.575-7.527-13.156-15.048-19.827-22.489-3.322-3.704-6.686-7.237-10.562-10.372-3.453-2.795-8.183-5.148-9.272-9.715-.307-1.29-.514-2.602-.853-3.883-.195-.74-.802-1.858-.707-2.664.19-1.6 3.15-2.211 4.347-2.753 4.143-1.875 8.334-3.732 12.398-5.774 2.245-1.128 4.24-2.503 6.33-3.877 1.758-1.158 3.763-1.507 5.746-2.122 4.404-1.366 8.469-3.313 12.533-5.465 4.195-2.22 8.425-4.44 12.491-6.887 4.129-2.485 7.332-5.54 10.832-8.785 1.565-1.45 2.678-.72 3.787.877 1.35 1.941 2.772 3.833 4.167 5.742.65.889 1.209 1.938 2.13 2.47.037.023.065.035.1.053-.16-.054-.33-.112-.522-.189.338 2.402 2.146 4.043 3.74 5.697-.15-.41-.292-.823-.434-1.236.399 1.144.95 1.92 1.686 2.895 1.464 1.94 3.111 3.648 4.842 5.351.895.88 1.8 1.752 2.705 2.625.36.348.724.697 1.084 1.048.583.564.912.745 1.523 1.02 1.238.559 1.526 1.535 2.594 2.306 1.605 1.16 3.603 1.926 5.265 3.07 1.356.933 2.536 2.83 4.105 3.177.568.125.99.034 1.447.483.602.59-.024.578.924.79.478.108.971-.166 1.453-.013.508.162.957.603 1.385.915.98.715 1.961 1.43 2.94 2.148a350.21 350.21 0 0 1 5.25 3.92c3.458 2.64 6.634 5.628 10.102 8.25 1.794 1.356 3.724 2.347 5.64 3.498 1.596.96 2.941 2.256 4.437 3.353 1.786 1.31 3.871 1.949 5.724 3.133 1.82 1.162 3.57 2.362 5.461 3.41 4.132 2.287 8.33 4.42 12.709 6.19 4.427 1.791 8.956 3.236 13.762 3.462 4.878.231 9.494-1.046 13.992-2.826a124.85 124.85 0 0 0 13.047-6.033c4.103-2.198 7.923-4.842 11.946-7.171a176.198 176.198 0 0 0 11.717-7.432c3.657-2.517 7.134-5.329 11.082-7.385 3.984-2.075 7.517-4.777 11.268-7.227 3.855-2.518 7.684-5.054 11.51-7.62 7.978-5.347 15.964-10.785 23.542-16.694 3.638-2.838 7.227-5.754 10.414-9.095 3.401-3.568 7.458-6.606 11.194-9.815l10.754-10.438a.174.174 0 0 0 .018-.018l-10.772 10.456c1.706-.32 2.815-1.428 4.185-2.49a231.124 231.124 0 0 0 4.289-3.406c2.063-1.68 4.45-3.407 4.21-6.383a48.817 48.817 0 0 1 2.786-2.434c1.57-1.252 3.442-2.016 5.11-3.125 3.425-2.276 6.365-5.174 9.582-7.719 1.416-1.118 2.687-1.965 2.908-3.884.24-2.083-.042-4.34-.056-6.439-.011-1.622-.476-3.886.144-5.421.623-1.539 2.693-2.199 4.033-2.933 3.784-2.076 7.511-4.252 11.298-6.323 3.727-2.038 6.967-3.396 11.23-3.76 1.855-.16 3.636-.439 5.471-.724 1.897-.296 3.664.264 5.522.604 4.213.773 8.477 1.234 12.689 2.017 4.241.788 8.436 1.837 12.625 2.87 4.193 1.033 8.378 2.133 12.635 2.88 2.178.381 4.43.632 6.541 1.312 1.597.513 3.132 1.237 4.774 1.6 3.975.876 8.015.98 11.9 2.308 3.879 1.325 7.836 1.993 11.88 2.601 2.117.32 4.227.66 6.322 1.113 1.264.274 2.388.435 3.67.121 2.656-.65 4.811-2.788 5.66-5.362 2.343-7.1-2.19-14.721-6.804-19.757-4.95-5.401-11.203-9.182-17.144-13.36 2.367-.26 4.72.089 7.086.132 2.027.037 4.011-.053 6.043.058 4.497.246 8.98-1.056 13.483-1.152 6.095-.13 13.616-1.373 15.208-8.32 1.481-6.465-5.377-12.292-10.27-15.226-3.144-1.885-6.327-3.61-9.406-5.605-3.29-2.131-7.355-3.385-10.989-4.819-3.698-1.46-7.422-2.887-11.27-3.904-3.697-.978-7.442-1.502-10.987-2.95 3.225-1.407 6.714-1.717 10.168-2.362 3.44-.643 6.821-1.494 10.192-2.426Z"
 					/>
